Changed to integer coordinates, and have reworked the logic behind the Agent class. Currently creating test cases for Agent, up to testing validDirection. Need to test for non % 16 coordinates, so I refactored the code to not have * 16 in the constructor. Checking for side effects.

+++ src/main/java/ghost/Utilities.java
@@ -4,7 +4,7 @@ import java.util.regex.*;
 import java.util.*;
 
 public class Utilities {
-	public static double distance(double[] a, double[] b) {
+	public static double distance(int[] a, int[] b) {
 		int xDist = (int)Math.pow(a[0] - b[0], 2);
 		int yDist = (int)Math.pow(a[1] - b[1], 2);
 		return (int)Math.pow(xDist + yDist, 0.5);
